Add PushResource and tests
This adds the final PushResource api endpoint and unit as well as
integration tests for it.

The integration test is disabled by default, as it needs a hardcoded
access token and we don't want to commit this to GitHub. The ideal
solution would be to hit a controlled server like WireMock or to read
a token from the env.